feat(build-ui) bug fixes
