+++ packages/build-ui/src/pages/agits/_id/management/artists/_id/components/artist_table.rs
@@ -116,7 +116,7 @@ pub fn ArtistTable(
                             validations::Extra { class: "[&>circle]:stroke-white", height: 18 }
                         }
                     }
-
+                
                 }
             }
             tbody {
@@ -189,7 +189,7 @@ pub fn ArtistTable(
                                             }
                                         }
                                     }
-
+                                
                                     td { class: "px-2 py-2 sm:px-4 sm:py-3 whitespace-nowrap",
                                         div { class: "flex flex-col",
                                             span { {asset.volume.to_string()} }
@@ -206,7 +206,7 @@ pub fn ArtistTable(
                                                 },
                                                 validations::Extra { class: "[&>circle]:stroke-white", height: 18 }
                                             }
-
+                                
                                             div {
                                                 class: "absolute right-0 mt-2 w-48 bg-black border border-green-500 rounded shadow-lg z-10 hidden aria-dropdown-open:block",
                                                 "aria-dropdown-open": is_dropdown_open,
@@ -237,7 +237,7 @@ pub fn ArtistTable(
                                                     }
                                                 }
                                             }
-
+                                        
                                         }
                                     }
                                 }
