+++ packages/build-ui/src/pages/agits/_id/management/artists/_id/components/input_field.rs
@@ -7,18 +7,18 @@ pub fn InputField(
     value: String,
 ) -> Element {
   rsx! {
-        div { class: "flex items-center",
-            label { class: "text-sm w-40",
-                span { class: "text-red-500 mr-1", "*" }
-                {label}
-            }
-            input {
-                class: "bg-border-background border border-border-primary text-white p-2.5 flex-1",
-                placeholder,
-                r#type: "text",
-                value,
-                oninput: move |e: Event<FormData>| onInput.call(e),
-            }
+    div { class: "flex items-center",
+        label { class: "text-sm w-40",
+            span { class: "text-red-500 mr-1", "*" }
+            {label}
+        }
+        input {
+            class: "bg-border-background border border-border-primary text-white p-2.5 flex-1",
+            placeholder,
+            r#type: "text",
+            value,
+            oninput: move |e: Event<FormData>| onInput.call(e),
         }
     }
+}
 }
