+++ packages/build-ui/src/pages/agits/_id/management/artists/_id/page.rs
@@ -91,7 +91,7 @@ pub fn ArtistDetailPage(lang: Language, agit_id: ReadOnlySignal<i64>, artist_id:
                     }
                 }
             }
-
+        
 
         }
     }
@@ -183,9 +183,9 @@ pub fn EditArtistPage(
                                     }
                                 }
                             }
-
+                        
                         }
-
+                    
 
 
 
@@ -299,7 +299,7 @@ pub fn EditArtistPage(
                         oninput: move |evt| {
                             ctrl.update_artist_field("introduction".to_string(), evt.value().clone())
                         },
-
+                    
                     }
                 }
