+++ packages/build-ui/src/pages/agits/_id/management/artists/new/page.rs
@@ -90,9 +90,9 @@ pub fn NewArtistPage(lang: Language, agit_id: ReadOnlySignal<i64>) -> Element {
                                     }
                                 }
                             }
-
+                        
                         }
-
+                    
 
 
 
@@ -204,7 +204,7 @@ pub fn NewArtistPage(lang: Language, agit_id: ReadOnlySignal<i64>) -> Element {
                         oninput: move |evt| {
                             ctrl.update_artist_field("introduction".to_string(), evt.value().clone())
                         },
-
+                    
                     }
                 }
