+++ packages/build-ui/src/pages/agits/_id/management/components/mod.rs
@@ -1,6 +1,9 @@
 
 mod nft_table;
 mod common_tables;
+mod table_head;
 
 pub use common_tables::{ActivityTable, OwnedTable};
 pub use nft_table::NftTable;
+pub use table_head::TableHeader;
+
